+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Last Changes
 
+
+## v2.0.0-beta.1
+
 - [#46](https://github.com/LaxarJS/laxar-mocks/issues/46): pass spec fixtures using the `init` method
     + **BREAKING CHANGE:** see ticket for details
 - [#45](https://github.com/LaxarJS/laxar-mocks/issues/45): publish `endLifecycleRequest` on tearDown